+++ packages/building-blocks/i18n/en-US.yml
@@ -1,0 +1,4 @@
+otpUi:
+  buildingBlocks:
+    alert:
+      expand: Expand
